% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possible_designs.R
\name{possible_designs}
\alias{possible_designs}
\title{possible_designs}
\usage{
possible_designs(design)
}
\arguments{
\item{design}{A named vector stating the which sample is assigned to which TMT channel.}
}
\value{
A data.frame containing all equivalent designs
}
\description{
Function to get all equivalent design producing the same global interference
}
