% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tmtinterference.R
\name{read_tmtinterference}
\alias{read_tmtinterference}
\title{read_tmtinterference}
\usage{
read_tmtinterference(f)
}
\arguments{
\item{f}{A path to a file containing the TMT reporter ion interference value.
Can be an xlsx, csv, txt tab delimited file or even the pdf product sheet.}
}
\value{
A data.frame containing the TMT reporter ion interference value
}
\description{
Function to read the TMT reporter ion interference value from a file
}
