% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runoptTMT.R
\name{runoptTMT}
\alias{runoptTMT}
\title{Run the app for optimizing your TMT experimental design}
\usage{
runoptTMT()
}
\description{
\code{runoptTMT} works exactly the same as runExample from \code{\link{shiny}} package.
}
