% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmt_interference_links.R
\name{tmt_interference_links}
\alias{tmt_interference_links}
\title{tmt_interference_links}
\usage{
tmt_interference_links(design, tmt_correction = NULL)
}
\arguments{
\item{design}{A named vector stating the which sample is assigned to which TMT channel.}

\item{tmt_correction}{A data.frame of three columns stating the amount of impurities in %
of the corresponding channels. If NULL, takes the default data from optTMT.}
}
\value{
A data.frame of two columns stating which channels interfere
}
\description{
Function to compute which channel will generate interference in your experimental design
}
