% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmt_interference_noise.R
\name{tmt_interference_noise}
\alias{tmt_interference_noise}
\title{tmt_interference_noise}
\usage{
tmt_interference_noise(design, tmt_correction = NULL)
}
\arguments{
\item{design}{A named vector stating the which sample is assigned to which TMT channel.}

\item{tmt_correction}{A data.frame of three columns stating the amount of impurities in %
of the corresponding channels. If NULL, takes the default data from optTMT.}
}
\value{
The global interference generated by your design in %
}
\description{
Function to compute the global interference generated by your design in %
}
